repair missing buffer initialization for bound vectors if column is NULL
This affects data types of:

- std::string
- std::wstring
- std::tm
- soci::binarydata

This seems to be a problem in debug versions but will also occure in rare
cases at release build. So the buffer has to be filled upfront with 0 to
ensure that scanning length (strlen/wcslen) doesn't fail while reading
column data and driver doesn't change the buffer.
It could also be handle during data assignment, but the indicator handling
have to be done upfront instead of 2nd case within post_fetch().